WIP: first commit of topK; moved summary and selective_MLE to query methods
